graph: correctly join multi event days into one data point
Every data point for the graph is supposed to be the bottom
line balance for that day, but currently, if there were multiple
balance events in a day, each of them will be considered its
own data point for the graph, causing a temporal shift in
the graph data.

Fix it by using only the latest balance event of each day as
that day's data point for the graph. This works because
if multiple events happened in a single day, the final log
event will represent the culmination of all the events.